# apps/api/openapi/compute.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoWaterSimu Compute API",
    "version": "0.1.0"
  },
  "servers": [{ "url": "http://localhost:8088" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/udm-models": {
      "get": {
        "operationId": "listUdmModels",
        "parameters": [
          { "name": "skip", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "UDM models", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListUDMModelsResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "createUdmModel",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Created UDM model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelDetailPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/udm-models/from-template": {
      "post": {
        "operationId": "createUdmModelFromTemplate",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelCreateFromTemplateRequest" } } }
        },
        "responses": {
          "201": { "description": "Created UDM model from template", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelDetailPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/udm-models/{model_id}": {
      "get": {
        "operationId": "getUdmModel",
        "parameters": [{ "name": "model_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "UDM model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelDetailPublic" } } } },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "patch": {
        "operationId": "updateUdmModel",
        "parameters": [{ "name": "model_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "Updated UDM model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelDetailPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "delete": {
        "operationId": "deleteUdmModel",
        "parameters": [{ "name": "model_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Archived UDM model", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/udm-hybrid-configs": {
      "get": {
        "operationId": "listUdmHybridConfigs",
        "parameters": [
          { "name": "skip", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } }
        ],
        "responses": {
          "200": { "description": "UDM hybrid configs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListUDMHybridConfigsResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "createUdmHybridConfig",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Created UDM hybrid config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "422": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/udm-hybrid-configs/{id}": {
      "get": {
        "operationId": "getUdmHybridConfig",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "UDM hybrid config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigPublic" } } } },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "patch": {
        "operationId": "updateUdmHybridConfig",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "Updated UDM hybrid config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" },
          "422": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "delete": {
        "operationId": "deleteUdmHybridConfig",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Archived UDM hybrid config", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
          "403": { "$ref": "#/components/responses/ContractError" },
          "404": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
...
    "responses": {
      "ContractError": { "description": "Contract error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContractError" } } } }
    },
    "schemas": {
...
      "UDMModelDefinitionDraft": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "components": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "parameters": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "processes": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "meta": { "type": "object", "additionalProperties": true },
          "validation_mode": { "type": "string" }
        }
      },
      "UDMModelCreateRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/UDMModelDefinitionDraft" },
          {
            "type": "object",
            "properties": {
              "seed_source": { "type": "string" },
              "source_system": { "type": "string" },
              "metadata": { "type": "object", "additionalProperties": true }
            }
          }
        ]
      },
      "UDMModelUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "components": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "parameters": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "processes": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "meta": { "type": "object", "additionalProperties": true },
          "is_published": { "type": "boolean" },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "UDMModelCreateFromTemplateRequest": {
        "type": "object",
        "required": ["template_key"],
        "properties": {
          "template_key": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
...
      "UDMModelPublic": {
        "type": "object",
        "required": ["id", "name", "tags", "current_version", "is_published", "owner_id", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "current_version": { "type": "integer" },
          "is_published": { "type": "boolean" },
          "owner_id": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "UDMModelVersionPublic": {
        "type": "object",
        "required": ["id", "model_id", "version", "content_hash", "components", "parameters", "processes", "validation_ok", "validation_errors", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string" },
          "model_id": { "type": "string" },
          "version": { "type": "integer" },
          "content_hash": { "type": "string" },
          "parameter_hash": { "type": "string" },
          "components": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "parameters": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "processes": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "meta": { "type": "object", "additionalProperties": true },
          "validation_ok": { "type": "boolean" },
          "validation_errors": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "seed_source": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "UDMModelDetailPublic": {
        "allOf": [
          { "$ref": "#/components/schemas/UDMModelPublic" },
          {
            "type": "object",
            "required": ["versions"],
            "properties": {
              "latest_version": { "$ref": "#/components/schemas/UDMModelVersionPublic" },
              "versions": { "type": "array", "items": { "$ref": "#/components/schemas/UDMModelVersionPublic" } }
            }
          }
        ]
      },
      "ListUDMModelsResponse": {
        "type": "object",
        "required": ["data", "count"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/UDMModelPublic" } },
          "count": { "type": "integer" }
        }
      },
      "UDMHybridConfigCreateRequest": {
        "type": "object",
        "required": ["name", "hybrid_config"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "hybrid_config": { "type": "object", "additionalProperties": true },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "UDMHybridConfigUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "hybrid_config": { "type": "object", "additionalProperties": true },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "UDMHybridConfigPublic": {
        "type": "object",
        "required": ["id", "owner_id", "created_at", "updated_at", "name", "hybrid_config"],
        "properties": {
          "id": { "type": "string" },
          "owner_id": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "hybrid_config": { "type": "object", "additionalProperties": true },
          "parameter_hash": { "type": "string" }
        }
      },
      "ListUDMHybridConfigsResponse": {
        "type": "object",
        "required": ["data", "count"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/UDMHybridConfigPublic" } },
          "count": { "type": "integer" }
        }
      },
      "ContractError": {
        "type": "object",
        "required": ["schema_version", "error_code", "message", "retryable"],
        "properties": {
          "schema_version": { "type": "string", "enum": ["contract_error.v1"] },
          "error_code": { "type": "string" },
          "message": { "type": "string" },
          "retryable": { "type": "boolean" },
          "details": { "type": "object", "additionalProperties": true }
        }
      },
...
    }
  }
}